requirements.txt
opencv-python
opencv-contrib-python
imutils
numpy
matplotlib
scikit-image

# test framework
Appium-Python-Client
pytest